--- emitter.js
@@ -1,3 +1,0 @@
-export default class Emitter {
-    constructor() {}
-}
